+++ bridge-core/src/lib.rs
@@ -4,7 +4,7 @@ pub mod transform;
 #[derive(Debug, PartialEq, Eq)]
 pub struct OffsetRecord {
     pub topic: String,
-    pub partition: i64,
+    pub partition: i32,
     pub offset: i64,
     pub consumer_group: String,
 }
